Update the IVtest
The new get_bias_curve method is implemented to obtain the vgate
curve from the minimum, maximum and step, which is read from the
/striptease/data/input_bias_IVtest.xlsx or from a input file.

The HK_SCAN is applied for polarimeters in all boards, which
retrieves the HK parameters for all polarimeter. This also can be
disabled.

The tags are reorder. For each parameter configuration, the tag
includes the Vgate and Vdrain values in the option "comment".

The log.info shows the configuration namber and estimation of
the total time of the test (in hrs and days).