Introduce new functions and LNA ordering
Reorder the number of the LNA (official name).

Provide the LNA list of one polarimeter.

Retrieve the drain voltage, gate voltage and drain current of an
amplifier.

Send a command to read and record the all House Keeping values for
all LNA of the polarimeters in a board (or boards).